renamed package of jcache sample infinispan-transaction-mode
